+++ axum/CHANGELOG.md
@@ -12,8 +12,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   `serde_json::Error` ([#1371])
 - **added**: `JsonRejection` now displays the path at which a deserialization
   error occurred too ([#1371])
+- **fixed:** Used `400 Bad Request` for `FailedToDeserializeQueryString`
+  rejections, instead of `422 Unprocessable Entity` ([#1387])
 
 [#1371]: https://github.com/tokio-rs/axum/pull/1371
+[#1387]: https://github.com/tokio-rs/axum/pull/1387
 
 # 0.6.0-rc.2 (10. September, 2022)